___utils___/requirements.txt
PyJWT==2.10.1
PyMySQL==1.1.1
annotated_types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.2.3
h11==0.16.0
idna==3.10
passlib==1.7.4
pip==25.1.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python_dotenv==1.1.0
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.46.2
typing_extensions==4.14.0
typing_inspection==0.4.1
uvicorn==0.34.3
